Create RandpassGenerator
I got this random password encrypter/decrypter to work. The only problem I am having now to see if it truly works, is that I cant get it to decrypt the password. File runs, create new folder if not in dir, creates password, encrypts and shows the encrypted key, but when it gets to the decrypt part it still prints but somethings not right. I think its still encrypted but slightly different then the original key. Any help to get this to work as intended would be greatly appreciated. Thank you